IGNITE-28938 Close the door behind the split: baseline lists, and a marshaller that no longer walks
Two messages still made their marshaller call MessageWire: the elements they rebuild from wire arrays were skipped by
the common walk, so the rebuild had to take them through the wire itself. Both halves of that are gone — the walk no
longer skips those arrays, and the rebuild only puts the elements back. No marshaller calls MessageWire now, so the
walking code moved to MessageWireFormGenerator, leaving the shared base at 281 lines from 526.

Codegen now checks two lists, custom-wire-form-messages.txt and marshallable-messages.txt: a message that takes part
in its own wiring by hand and is not listed is rejected at compile time. The lists name who does it today. They are
meant to shrink, and when the wire form one empties, the deprecated interface goes away with the step that runs it.
