+++ modules/codegen/src/main/java/org/apache/ignite/internal/MessageWireFormGenerator.java
@@ -21,13 +21,19 @@ import java.io.IOException;
 import java.io.StringWriter;
 import java.io.Writer;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 import java.util.Set;
 import javax.annotation.processing.ProcessingEnvironment;
 import javax.lang.model.element.Element;
+import javax.lang.model.element.QualifiedNameable;
 import javax.lang.model.element.TypeElement;
 import javax.lang.model.element.VariableElement;
+import javax.lang.model.type.ArrayType;
+import javax.lang.model.type.DeclaredType;
+import javax.lang.model.type.TypeKind;
 import javax.lang.model.type.TypeMirror;
+import javax.lang.model.type.TypeVariable;
 import javax.lang.model.util.ElementFilter;
 import javax.tools.Diagnostic;
 import org.apache.ignite.internal.systemview.SystemViewRowAttributeWalkerProcessor;
@@ -81,6 +87,45 @@ public class MessageWireFormGenerator extends MessageWireCompanionGenerator {
 
             return writer.toString();
         }
+    }
+
+    /** Cache-free read-back of a {@code @NioField} message field on the NIO thread (no cache context available). */
+    private List<String> fromWireNioField(String accessor) {
+        imports.add(MESSAGE_WIRE_CLS);
+
+        List<String> code = new ArrayList<>();
+
+        code.add(indentedLine("if (%s != null)", accessor));
+
+        indent++;
+
+        code.add(indentedLine("MessageWire.fromWire(%s, kctx);", accessor));
+
+        indent--;
+
+        return code;
+    }
+
+    /**
+     * Returns whether the {@code @Order} fields of {@code msgType} need a cache object context to unmarshal. Such a
+     * message must not be a {@code @NioField}: its {@code unmarshalNio} runs on the NIO thread, which has no context.
+     * A type with no fields to inspect (e.g. a type variable) is conservatively assumed to need the context.
+     */
+    private boolean nestedNeedsCtx(TypeMirror type) {
+        Element el = env.getTypeUtils().asElement(type);
+
+        if (!(el instanceof TypeElement))
+            return true;
+
+        return SystemViewRowAttributeWalkerProcessor.superclasses(env, (TypeElement)el)
+            .flatMap(c -> ElementFilter.fieldsIn(c.getEnclosedElements()).stream())
+            .filter(f -> f.getAnnotation(Order.class) != null)
+            .anyMatch(f -> needsCtxType(f.asType()));
+    }
+
+    /** */
+    private static boolean isNioField(VariableElement field) {
+        return field.getAnnotation(NioField.class) != null;
     }
 
     /** Generates the {@code toWire} method: every field on the way out. */
@@ -148,11 +193,9 @@ public class MessageWireFormGenerator extends MessageWireCompanionGenerator {
         hasStatements |= emitMethod(methods, "fromWire(" + params + ")", body -> {
             usesMsgFactory = false;
 
-            Set<String> wireFieldSkip = marshalledWireFieldsToSkip();
-
             List<String> code = new ArrayList<>();
 
-            appendFields(code, fields, Direction.IN, wireFieldSkip);
+            appendFields(code, fields, Direction.IN);
 
             if (code.isEmpty())
                 return;
@@ -174,8 +217,54 @@ public class MessageWireFormGenerator extends MessageWireCompanionGenerator {
         });
     }
 
-    /** Cache-free read-back of a {@code @NioField} message field on the NIO thread (no cache context available). */
-    private List<String> fromWireNioField(String accessor) {
+    /** Generates the code of each field and appends the non-empty results to {@code body}. */
+    private void appendFields(List<String> body, List<VariableElement> fields, Direction mode) {
+        appendFields(body, fields, mode, Set.of());
+    }
+
+    /** Generates the code of each field, skipping names in {@code skip}, and appends the non-empty results. */
+    private void appendFields(List<String> body, List<VariableElement> fields, Direction mode, Set<String> skip) {
+        for (VariableElement field : fields) {
+            if (skip.contains(field.getSimpleName().toString()))
+                continue;
+
+            List<String> result = codeFor(field.asType(), fieldAccessor(field), mode);
+
+            if (!result.isEmpty())
+                appendBlock(body, result);
+        }
+    }
+
+    /**
+     * @return the generated code lines for a field of type {@code t} in the given direction, or empty if it needs none.
+     */
+    private List<String> codeFor(TypeMirror t, String accessor, Direction mode) {
+        if (t.getKind() == TypeKind.ARRAY) {
+            TypeMirror comp = ((ArrayType)t).getComponentType();
+
+            return comp.getKind() == TypeKind.DECLARED ? arrayCode(comp, accessor, mode) : List.of();
+        }
+
+        if (t.getKind() == TypeKind.DECLARED || t.getKind() == TypeKind.TYPEVAR) {
+            if (isMessage(t))
+                return isNonMarshallable(t) ? List.of() : messageCode(accessor, mode);
+            if (isCacheObject(t))
+                return cacheObjectCode(accessor, mode);
+            if (isMap(t))
+                return mapCode((DeclaredType)t, accessor, mode);
+            if (isCollection(t))
+                return collectionCode((DeclaredType)t, accessor, mode);
+        }
+
+        return List.of();
+    }
+
+    /**
+     * Generates a null-guarded {@code MessageWire} call. Loop-nested calls go through the overloads taking the
+     * pre-resolved {@code msgFactory} local (see {@link #prependMsgFactoryResolution}), so the factory is not
+     * re-resolved from the context on every element.
+     */
+    private List<String> messageCode(String accessor, Direction mode) {
         imports.add(MESSAGE_WIRE_CLS);
 
         List<String> code = new ArrayList<>();
@@ -184,32 +273,173 @@ public class MessageWireFormGenerator extends MessageWireCompanionGenerator {
 
         indent++;
 
-        code.add(indentedLine("MessageWire.fromWire(%s, kctx);", accessor));
+        if (loopDepth > 0) {
+            usesMsgFactory = true;
+
+            code.add(mode == Direction.OUT
+                ? indentedLine("MessageWire.toWire(msgFactory, %s, kctx, ctx);", accessor)
+                : indentedLine("MessageWire.fromWire(msgFactory, %s, kctx, ctx, clsLdr);", accessor));
+        }
+        else {
+            code.add(mode == Direction.OUT
+                ? indentedLine("MessageWire.toWire(%s, kctx, ctx);", accessor)
+                : indentedLine("MessageWire.fromWire(%s, kctx, ctx, clsLdr);", accessor));
+        }
 
         indent--;
 
         return code;
     }
 
-    /** */
-    private static boolean isNioField(VariableElement field) {
-        return field.getAnnotation(NioField.class) != null;
+    /** Generates a null-and-ctx-guarded call that prepares a {@code CacheObject} field, or reads it back. */
+    private List<String> cacheObjectCode(String accessor, Direction mode) {
+        List<String> code = new ArrayList<>();
+
+        code.add(indentedLine("if (%s != null && ctx != null)", accessor));
+
+        indent++;
+
+        code.add(mode == Direction.OUT
+            ? indentedLine("%s.marshal(ctx);", accessor)
+            : indentedLine("%s.unmarshal(ctx, clsLdr);", accessor));
+
+        indent--;
+
+        return code;
     }
 
-    /**
-     * Returns whether the {@code @Order} fields of {@code msgType} need a cache object context to unmarshal. Such a
-     * message must not be a {@code @NioField}: its {@code unmarshalNio} runs on the NIO thread, which has no context.
-     * A type with no fields to inspect (e.g. a type variable) is conservatively assumed to need the context.
-     */
-    private boolean nestedNeedsCtx(TypeMirror type) {
-        Element el = env.getTypeUtils().asElement(type);
+    /** Generates a null-guarded for-each loop over the array's elements. */
+    private List<String> arrayCode(TypeMirror comp, String accessor, Direction mode) {
+        Element elem = ((DeclaredType)comp).asElement();
 
-        if (!(el instanceof TypeElement))
-            return true;
+        indent++;
 
-        return SystemViewRowAttributeWalkerProcessor.superclasses(env, (TypeElement)el)
-            .flatMap(c -> ElementFilter.fieldsIn(c.getEnclosedElements()).stream())
-            .filter(f -> f.getAnnotation(Order.class) != null)
-            .anyMatch(f -> needsCtxType(f.asType()));
+        List<String> loopCode = forLoop(elem.getSimpleName().toString(), comp, accessor, mode);
+
+        indent--;
+
+        if (!loopCode.isEmpty())
+            imports.add(((QualifiedNameable)elem).getQualifiedName().toString());
+
+        return wrapNullGuarded(accessor, loopCode);
+    }
+
+    /** Generates a null-guarded for-each loop over the collection's elements. */
+    private List<String> collectionCode(DeclaredType t, String accessor, Direction mode) {
+        TypeMirror arg = t.getTypeArguments().get(0);
+
+        if (arg.getKind() != TypeKind.DECLARED && arg.getKind() != TypeKind.TYPEVAR)
+            return List.of();
+
+        Element elem = element(arg);
+
+        String typeName = elem.getSimpleName().toString();
+
+        indent++;
+
+        List<String> loopCode = forLoop(typeName, arg, "(Collection<? extends " + typeName + ">)" + accessor, mode);
+
+        indent--;
+
+        if (!loopCode.isEmpty()) {
+            imports.add(((QualifiedNameable)elem).getQualifiedName().toString());
+            imports.add(Collection.class.getName());
+        }
+
+        return wrapNullGuarded(accessor, loopCode);
+    }
+
+    /** Iterates {@code keySet()} then {@code values()}, wrapping both loops in a null-guard. */
+    private List<String> mapCode(DeclaredType t, String accessor, Direction mode) {
+        List<? extends TypeMirror> args = t.getTypeArguments();
+
+        indent++;
+
+        List<String> combined = new ArrayList<>();
+
+        for (int i = 0; i < 2; i++) {
+            TypeMirror elemType = args.get(i);
+
+            if (elemType.getKind() != TypeKind.DECLARED && elemType.getKind() != TypeKind.TYPEVAR)
+                continue;
+
+            Element elem = element(elemType);
+
+            String typeName = elem.getSimpleName().toString();
+            String collection = i == 0 ? "keySet" : "values";
+            String iterable = "((Collection<? extends " + typeName + ">)" + accessor + "." + collection + "())";
+
+            List<String> loopCode = forLoop(typeName, elemType, iterable, mode);
+
+            if (loopCode.isEmpty())
+                continue;
+
+            imports.add(((QualifiedNameable)elem).getQualifiedName().toString());
+            imports.add(Collection.class.getName());
+
+            combined.addAll(loopCode);
+        }
+
+        indent--;
+
+        return wrapNullGuarded(accessor, combined);
+    }
+
+    /** @return a for-each loop over {@code iterable}, or empty when its elements need no code of their own. */
+    private List<String> forLoop(String typeName, TypeMirror elemType, String iterable, Direction mode) {
+        String el = loopDepth == 0 ? "e" : "e" + loopDepth;
+
+        loopDepth++;
+        indent++;
+
+        List<String> inner = codeFor(elemType, el, mode);
+
+        indent--;
+        loopDepth--;
+
+        if (inner.isEmpty())
+            return List.of();
+
+        List<String> code = new ArrayList<>();
+
+        code.add(indentedLine("for (%s %s : %s) {", typeName, el, iterable));
+
+        code.addAll(inner);
+
+        code.add(indentedLine("}"));
+
+        return code;
+    }
+
+    /** Returns empty if {@code inner} is empty; otherwise wraps {@code inner} in a null-guard on {@code nullGuard}. */
+    private List<String> wrapNullGuarded(String nullGuard, List<String> inner) {
+        if (inner.isEmpty())
+            return List.of();
+
+        List<String> code = new ArrayList<>();
+
+        code.add(indentedLine("if (%s != null) {", nullGuard));
+
+        code.addAll(inner);
+
+        code.add(indentedLine("}"));
+
+        return code;
+    }
+
+    /** Which way the generated field code runs: {@link #OUT} before the message is written, {@link #IN} after it is read. */
+    private enum Direction {
+        /** On the way out: the code runs before the message is written. */
+        OUT,
+
+        /** On the way in: the code runs after the message is read, with cache context and class loader at hand. */
+        IN
+    }
+
+    /** Returns the element for {@code t}; for a type variable, uses its upper bound. */
+    private Element element(TypeMirror t) {
+        return t.getKind() == TypeKind.DECLARED
+            ? ((DeclaredType)t).asElement()
+            : ((DeclaredType)((TypeVariable)t).getUpperBound()).asElement();
     }
 }
