+++ modules/codegen/src/main/resources/marshallable-messages.txt
@@ -1,0 +1,18 @@
+# Messages allowed to implement MarshallableMessage, that is, to turn some of their fields into bytes themselves.
+# Codegen does this for a @Marshalled field on its own, so a hand-written marshal() is the exception, not the rule.
+# This list only shrinks. Adding a message here needs a reason: marshal() must produce bytes, not move fields around
+# — moving fields is the wire form, and that one is deprecated.
+#
+# A message not listed here is rejected at compile time.
+org.apache.ignite.internal.GridTopicMessage
+org.apache.ignite.internal.managers.communication.ErrorMessage
+org.apache.ignite.internal.managers.eventstorage.GridEventStorageMessage
+org.apache.ignite.internal.processors.cache.distributed.dht.atomic.GridNearAtomicFullUpdateRequest
+org.apache.ignite.internal.processors.cache.distributed.dht.atomic.GridNearAtomicSingleUpdateInvokeRequest
+org.apache.ignite.internal.processors.cache.transactions.IgniteTxEntry
+org.apache.ignite.internal.processors.query.calcite.message.CalciteErrorMessage
+org.apache.ignite.spi.discovery.SerializableDataBagItemWrapper
+
+# Fixtures of the tests: they exist to prove the mechanism still works.
+org.apache.ignite.internal.TestMarshallableMessage
+org.apache.ignite.spi.discovery.tcp.ExploitMessage
